requirements.txt
FastAPI[standard]
BeaultifulSoup4
python-whois
requests
pyOpenSSL